requirements.txt
python==3.8
astroid==2.3.3
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
Flask==1.1.2
fonttools==4.25.0
idna==3.2
importlib-metadata==4.8.1
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numexpr==2.7.3
numpy==1.20.3
olefile==0.46
pandas==1.0.3
Pillow==8.3.1
pip==21.2.2
pycparser==2.20
pylint==2.4.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2019.3
requests==2.26.0
setuptools==58.0.4
sip==4.19.13
six==1.16.0
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.5.0
gunicorn

